Collect reply markup stuff in one place
