+++ inline.d.ts
@@ -1,4 +1,5 @@
 import { Chat, User } from "./manage";
+import { InlineKeyboardMarkup } from "./markup";
 import { Location, MessageEntity, ParseMode } from "./message";
 import { LabeledPrice } from "./payment";
 
@@ -17,84 +18,6 @@ export interface InlineQuery {
   /** Sender location, only for bots that request user location */
   location?: Location;
 }
-
-/** This object represents an inline keyboard that appears right next to the message it belongs to. */
-export interface InlineKeyboardMarkup {
-  /** Array of button rows, each represented by an Array of InlineKeyboardButton objects */
-  inline_keyboard: InlineKeyboardButton[][];
-}
-
-export namespace InlineKeyboardButton {
-  interface AbstractInlineKeyboardButton {
-    /** Label text on the button */
-    text: string;
-  }
-  export interface UrlButton extends AbstractInlineKeyboardButton {
-    /** HTTP or tg:// url to be opened when the button is pressed. Links tg://user?id=<user_id> can be used to mention a user by their ID without using a username, if this is allowed by their privacy settings. */
-    url: string;
-  }
-  export interface LoginButton extends AbstractInlineKeyboardButton {
-    /** An HTTP URL used to automatically authorize the user. Can be used as a replacement for the Telegram Login Widget. */
-    login_url: LoginUrl;
-  }
-  export interface CallbackButton extends AbstractInlineKeyboardButton {
-    /** Data to be sent in a callback query to the bot when button is pressed, 1-64 bytes */
-    callback_data: string;
-  }
-  export interface SwitchInlineButton extends AbstractInlineKeyboardButton {
-    /** If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified inline query in the input field. Can be empty, in which case just the bot's username will be inserted.
-
-    Note: This offers an easy way for users to start using your bot in inline mode when they are currently in a private chat with it. Especially useful when combined with switch_pm… actions – in this case the user will be automatically returned to the chat they switched from, skipping the chat selection screen. */
-    switch_inline_query: string;
-  }
-  export interface SwitchInlineCurrentChatButton
-    extends AbstractInlineKeyboardButton {
-    /** If set, pressing the button will insert the bot's username and the specified inline query in the current chat's input field. Can be empty, in which case only the bot's username will be inserted.
-
-    This offers a quick way for the user to open your bot in inline mode in the same chat – good for selecting something from multiple options. */
-    switch_inline_query_current_chat: string;
-  }
-  export interface GameButton extends AbstractInlineKeyboardButton {
-    /** Description of the game that will be launched when the user presses the button.
-
-    NOTE: This type of button must always be the first button in the first row. */
-    callback_game: CallbackGame;
-  }
-  export interface PayButton extends AbstractInlineKeyboardButton {
-    /** Specify True, to send a Pay button.
-
-    NOTE: This type of button must always be the first button in the first row and can only be used in invoice messages. */
-    pay: boolean;
-  }
-}
-
-/** This object represents one button of an inline keyboard. You must use exactly one of the optional fields. */
-export type InlineKeyboardButton =
-  | InlineKeyboardButton.CallbackButton
-  | InlineKeyboardButton.GameButton
-  | InlineKeyboardButton.LoginButton
-  | InlineKeyboardButton.PayButton
-  | InlineKeyboardButton.SwitchInlineButton
-  | InlineKeyboardButton.SwitchInlineCurrentChatButton
-  | InlineKeyboardButton.UrlButton;
-
-/** This object represents a parameter of the inline keyboard button used to automatically authorize a user. Serves as a great replacement for the Telegram Login Widget when the user is coming from Telegram. All the user needs to do is tap/click a button and confirm that they want to log in.
-Telegram apps support these buttons as of version 5.7. */
-export interface LoginUrl {
-  /** An HTTP URL to be opened with user authorization data added to the query string when the button is pressed. If the user refuses to provide authorization data, the original URL without information about the user will be opened. The data added is the same as described in Receiving authorization data.
-
-  NOTE: You must always check the hash of the received data to verify the authentication and the integrity of the data as described in Checking authorization. */
-  url: string;
-  /** New text of the button in forwarded messages. */
-  forward_text?: string;
-  /** Username of a bot, which will be used for user authorization. See Setting up a bot for more details. If not specified, the current bot's username will be assumed. The url's domain must be the same as the domain linked with the bot. See Linking your domain to the bot for more details. */
-  bot_username?: string;
-  /** Pass True to request the permission for your bot to send messages to the user. */
-  request_write_access?: boolean;
-}
-
-/** A placeholder, currently holds no information. Use BotFather to set up your game. */
-export interface CallbackGame {}
 
 /** This object represents one result of an inline query. Telegram clients currently support results of the following 20 types:
 - InlineQueryResultCachedAudio
