+++ proxied.d.ts
@@ -2,10 +2,11 @@ import { ApiResponse } from "./api";
 import { BotCommandScope } from "./bot-command-scope";
 import {
   ForceReply,
+  InlineKeyboardMarkup,
   ReplyKeyboardMarkup,
   ReplyKeyboardRemove,
-} from "./callback";
-import { InlineKeyboardMarkup, InlineQueryResult } from "./inline";
+} from "./markup";
+import { InlineQueryResult } from "./inline";
 import {
   BotCommand,
   ChatFromGetChat,
